Volume 5 db decrease added. lower_volume_in_segments moved to separate file.
